Improve tutorials build, move computation driver (toDot) to tutorials
