+++ src/CMakeLists.txt
@@ -56,9 +56,6 @@ list(REMOVE_ITEM iegenlib_SOURCES
 list(REMOVE_ITEM iegenlib_SOURCES
   ${IEGENLIB_SOURCE_DIR}/src/drivers/subSetDriver.cc)
 
-list(REMOVE_ITEM iegenlib_SOURCES
-  ${IEGENLIB_SOURCE_DIR}/src/drivers/Computation_driver.cc)
-
 
 
 ############################### STEP 2 ########################################
@@ -134,11 +131,6 @@ cmake_policy(SET CMP0037 OLD)
 #Compile and link the subSetDriver executable
 add_executable(../bin/subSetDriver drivers/subSetDriver.cc)
 target_link_libraries(../bin/subSetDriver iegenlib isl codegen omega gmp)
-
-cmake_policy(SET CMP0037 OLD)
-#Compile and link the Computation executable
-add_executable(../bin/Computation_driver drivers/Computation_driver.cc)
-target_link_libraries(../bin/Computation_driver iegenlib isl codegen omega gmp)
